+++ hnnt-chat-web/src/screens/Messaging/Messaging.js
@@ -3,6 +3,7 @@ import TabChatRightBar from '../../components/Tab/TabChatRightBar';
 import TabChatLeftBar from '../../components/Tab/TabChatLeftBar';
 import TabChatRightBarSearch from '../../components/Tab/TabChatRightBarSearch';
 import { useSelector } from 'react-redux';
+import TabMessage from '../../components/Tab/TabMessage';
 
 function Messaging() {
     const userActive = useSelector((state) => state.auth.userActive);
@@ -21,7 +22,7 @@ function Messaging() {
                     }`}
                 >
                     {!activeChat?.delete.includes(userId) && activeChat ? (
-                        <TabChat />
+                        <TabMessage />
                     ) : (
                         <div className="flex-1 flex items-center justify-center text-gray-500 dark:text-white">
                             Chọn một cuộc trò chuyện để bắt đầu
